Braeden - Created Observers and more fully fleshed out the scoreModel class
Created Observers and more fully fleshed out the scoreModel class. I created a mutable live data variable that allows for a change in the number of moves to be updated and track consistently and automatically. I also have created ViewModelProviders that are being grabbed from the scoreModel class and placed ViewModelProviders in every activity that contains a move counter, allowing access to the scoreModel class and its attributes.