Braeden - Implemented ScoreScreenActivity Functionality
Implemented ScoreScreenActivity Functionality. The puzzle activity, once completed, will send extras alongside the intent to the ScoreScreenActivity that will allows for the activity to determine which thresholds need to be crossed in order to get the corresponding amount of stars. Added in star fade in animation, left some commented out code inside the activity in case of revision later. I also added color codes to the colors.xml file